+++ src/main/java/io/swagger/codegen/v3/generators/kotlin/KotlinServerCodegen.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.languages.kotlin;
+package io.swagger.codegen.v3.generators.kotlin;
 
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.OpenAPI;
@@ -10,7 +10,6 @@ import io.swagger.codegen.v3.CliOption;
 import io.swagger.codegen.v3.CodegenConstants;
 import io.swagger.codegen.v3.CodegenType;
 import io.swagger.codegen.v3.CodegenModel;
-import io.swagger.codegen.v3.CodegenProperty;
 import io.swagger.codegen.v3.CodegenOperation;
 import io.swagger.codegen.v3.SupportingFile;
 import org.slf4j.Logger;
