+++ src/main/java/io/swagger/codegen/v3/generators/php/PhpClientCodegen.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.languages.php;
+package io.swagger.codegen.v3.generators.php;
 
 import io.swagger.codegen.v3.CliOption;
 import io.swagger.codegen.v3.CodegenConstants;
@@ -7,7 +7,7 @@ import io.swagger.codegen.v3.CodegenParameter;
 import io.swagger.codegen.v3.CodegenProperty;
 import io.swagger.codegen.v3.CodegenType;
 import io.swagger.codegen.v3.SupportingFile;
-import io.swagger.codegen.languages.DefaultCodegenConfig;
+import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
 
 import java.io.File;
 import java.util.Arrays;
