+++ src/main/java/io/swagger/codegen/v3/generators/swift/AbstractSwiftCodegen.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.languages.swift;
+package io.swagger.codegen.v3.generators.swift;
 
 
 import io.swagger.codegen.v3.CodegenConstants;
@@ -6,7 +6,7 @@ import io.swagger.codegen.v3.CodegenModel;
 import io.swagger.codegen.v3.CodegenProperty;
 import io.swagger.codegen.v3.CodegenType;
 import io.swagger.codegen.v3.SupportingFile;
-import io.swagger.codegen.languages.DefaultCodegenConfig;
+import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.MapSchema;
 import io.swagger.v3.oas.models.media.Schema;
