+++ src/main/java/io/swagger/codegen/v3/generators/swift/Swift3Codegen.java
@@ -1,7 +1,8 @@
-package io.swagger.codegen.languages.swift;
+package io.swagger.codegen.v3.generators.swift;
 
 import io.swagger.codegen.v3.CliOption;
 import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.generators.swift.AbstractSwiftCodegen;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
