+++ src/main/resources/META-INF/services/io.swagger.codegen.CodegenConfig
@@ -1,20 +1,20 @@
 io.swagger.codegen.v3.generators.html.StaticDocCodegen
 io.swagger.codegen.v3.generators.html.StaticHtmlCodegen
 io.swagger.codegen.v3.generators.html.StaticHtml2Codegen
-io.swagger.codegen.languages.java.JavaClientCodegen
-io.swagger.codegen.languages.java.JavaCXFClientCodegen
-io.swagger.codegen.languages.java.JavaCXFServerCodegen
-io.swagger.codegen.languages.java.JavaInflectorServerCodegen
-io.swagger.codegen.languages.java.JavaJAXRSCXFCDIServerCodegen
-io.swagger.codegen.languages.java.JavaJAXRSSpecServerCodegen
-io.swagger.codegen.languages.java.JavaJerseyServerCodegen
-io.swagger.codegen.languages.java.JavaJerseyDIServerCodegen
-io.swagger.codegen.languages.java.JavaResteasyEapServerCodegen
-io.swagger.codegen.languages.java.JavaResteasyServerCodegen
-io.swagger.codegen.languages.java.SpringCodegen
-io.swagger.codegen.languages.kotlin.KotlinClientCodegen
-io.swagger.codegen.languages.kotlin.KotlinServerCodegen
-io.swagger.codegen.languages.php.PhpClientCodegen
-io.swagger.codegen.languages.swift.Swift3Codegen
-io.swagger.codegen.languages.swift.Swift4Codegen
-io.swagger.codegen.languages.typescript.TypeScriptAngularClientCodegen
+io.swagger.codegen.v3.generators.java.JavaClientCodegen
+io.swagger.codegen.v3.generators.java.JavaCXFClientCodegen
+io.swagger.codegen.v3.generators.java.JavaCXFServerCodegen
+io.swagger.codegen.v3.generators.java.JavaInflectorServerCodegen
+io.swagger.codegen.v3.generators.java.JavaJAXRSCXFCDIServerCodegen
+io.swagger.codegen.v3.generators.java.JavaJAXRSSpecServerCodegen
+io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen
+io.swagger.codegen.v3.generators.java.JavaJerseyDIServerCodegen
+io.swagger.codegen.v3.generators.java.JavaResteasyEapServerCodegen
+io.swagger.codegen.v3.generators.java.JavaResteasyServerCodegen
+io.swagger.codegen.v3.generators.java.SpringCodegen
+io.swagger.codegen.v3.generators.kotlin.KotlinClientCodegen
+io.swagger.codegen.v3.generators.kotlin.KotlinServerCodegen
+io.swagger.codegen.v3.generators.php.PhpClientCodegen
+io.swagger.codegen.v3.generators.swift.Swift3Codegen
+io.swagger.codegen.v3.generators.swift.Swift4Codegen
+io.swagger.codegen.v3.generators.typescript.TypeScriptAngularClientCodegen
