+++ src/main/resources/v2/codegen/generatorClass.mustache
@@ -1,7 +1,7 @@
 package {{generatorPackage}};
 
 import io.swagger.codegen.*;
-import io.swagger.codegen.languages.DefaultCodegenConfig;
+import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
 
 import java.util.*;
 import java.io.File;
