+++ src/test/java/io/swagger/codegen/languages/DefaultCodegenConfigTest.java
@@ -4,6 +4,7 @@ import io.swagger.codegen.CodegenArgument;
 import io.swagger.codegen.CodegenConstants;
 import io.swagger.codegen.CodegenProperty;
 import io.swagger.codegen.CodegenType;
+import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
 import io.swagger.v3.oas.models.media.NumberSchema;
 import io.swagger.v3.oas.models.media.Schema;
 import org.testng.Assert;
