+++ src/test/java/io/swagger/codegen/languages/java/AbstractJavaCodegenTest.java
@@ -3,6 +3,7 @@ package io.swagger.codegen.languages.java;
 import io.swagger.codegen.CodegenArgument;
 import io.swagger.codegen.CodegenConstants;
 import io.swagger.codegen.CodegenType;
+import io.swagger.codegen.v3.generators.java.AbstractJavaCodegen;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.PathItem;
