+++ src/test/java/io/swagger/codegen/languages/swift/Swift3CodegenTest.java
@@ -2,14 +2,12 @@ package io.swagger.codegen.languages.swift;
 
 import io.swagger.codegen.CodegenConstants;
 import io.swagger.codegen.CodegenOperation;
-import io.swagger.codegen.languages.java.JavaClientCodegen;
+import io.swagger.codegen.v3.generators.swift.Swift3Codegen;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.parser.OpenAPIV3Parser;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-
-import java.util.Collections;
 
 import static io.swagger.codegen.handlebars.helpers.ExtensionHelper.getBooleanValue;
 
@@ -84,7 +82,7 @@ public class Swift3CodegenTest {
 
         Assert.assertEquals(op.returnType, "Data");
         // TODO - have to be checked. It seems that there is a new expected value "Object" set in the method:
-        // io.swagger.codegen.languages.DefaultCodegenConfig.fromRequestBody
+        // io.swagger.codegen.v3.generators.DefaultCodegenConfig.fromRequestBody
 //        Assert.assertEquals(op.bodyParam.dataType, "Data");
         Assert.assertTrue(getBooleanValue(op.bodyParam, CodegenConstants.IS_BINARY_EXT_NAME));
