+++ src/test/java/io/swagger/codegen/languages/swift/Swift4ModelTest.java
@@ -3,6 +3,7 @@ package io.swagger.codegen.languages.swift;
 import io.swagger.codegen.CodegenConstants;
 import io.swagger.codegen.CodegenModel;
 import io.swagger.codegen.CodegenProperty;
+import io.swagger.codegen.v3.generators.swift.Swift4Codegen;
 import io.swagger.v3.oas.models.media.BinarySchema;
 import io.swagger.v3.oas.models.media.ByteArraySchema;
 import io.swagger.v3.oas.models.media.DateSchema;
