+++ src/test/java/io/swagger/codegen/options/JavaOptionsProvider.java
@@ -2,7 +2,7 @@ package io.swagger.codegen.options;
 
 import com.google.common.collect.ImmutableMap;
 import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.languages.java.JavaClientCodegen;
+import io.swagger.codegen.v3.generators.java.JavaClientCodegen;
 
 import java.util.Map;
